Problem: synchronizing startup of orchestrator & parity
At the present moment if orchestrator fails to establish connection to
parity or callback to 30303 is unsuccessful, it will immediately stop.
When started, parity takes some time to become fully operational and
sync with the network, furthermore, the moment rpc is ready cannot be
reliably tracked from outside, so it makes it difficult to automate
pairing orchestrator with parity.

Solution: keep re-trying connection to Parity (both local
and external) until it succeeds.

Closes #2